Refactor - Extract GenerateMigrationReport logic into service
This introduces a seam, allowing us to test without worrying about the azure functions trigger parameters.
